+++ source/Application/Controller/Admin/ThemeMain.php
@@ -6,6 +6,7 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
+use OxidEsales\Eshop\Core\Registry;
 use oxRegistry;
 use oxTheme;
 use oxException;
@@ -81,9 +82,9 @@ class ThemeMain extends \OxidEsales\Eshop\Application\Controller\Admin\AdminDeta
         try {
             $oTheme->activate();
             $this->resetContentCache();
-        } catch (\OxidEsales\Eshop\Core\Exception\StandardException $oEx) {
-            \OxidEsales\Eshop\Core\Registry::getUtilsView()->addErrorToDisplay($oEx);
-            $oEx->debugOut();
+        } catch (\OxidEsales\Eshop\Core\Exception\StandardException $exception) {
+            Registry::getUtilsView()->addErrorToDisplay($exception);
+            Registry::getLogger()->error($exception->getMessage(), [$exception]);
         }
     }
 }