+++ source/Application/Model/Shop.php
@@ -6,6 +6,7 @@
 
 namespace OxidEsales\EshopCommunity\Application\Model;
 
+use OxidEsales\Eshop\Core\Registry;
 use oxRegistry;
 use oxDb;
 
@@ -359,7 +360,7 @@ class Shop extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel
             try {
                 $oDb->execute($sQuery);
             } catch (\OxidEsales\Eshop\Core\Exception\StandardException $exception) {
-                $exception->debugOut();
+                \OxidEsales\Eshop\Core\Registry::getLogger()->error($exception->getMessage(), [$exception]);
                 $bSuccess = false;
             }
         }