+++ source/Core/OnlineCaller.php
@@ -93,7 +93,8 @@ abstract class OnlineCaller
             $this->_resetFailedCallsCount($iFailedCallsCount);
         } catch (Exception $oEx) {
             if ($iFailedCallsCount > self::ALLOWED_HTTP_FAILED_CALLS_COUNT) {
-                $this->_castExceptionAndWriteToLog($oEx);
+                \OxidEsales\Eshop\Core\Registry::getLogger()->error($oEx->getMessage(), [$oEx]);
+
                 $sXml = $this->_formEmail($oRequest);
                 $this->_sendEmail($sXml);
                 $this->_resetFailedCallsCount($iFailedCallsCount);
@@ -103,24 +104,6 @@ abstract class OnlineCaller
         }
 
         return $sOutputXml;
-    }
-
-    /**
-     * Depending on the type of exception, first cast the exception and then write it to log.
-     *
-     * @deprecated since v6.3 (2018-04-25); This method will be removed completely. Use Registry::getLogger() to log error messages in the future.
-     *
-     * @param \Exception $oEx
-     */
-    protected function _castExceptionAndWriteToLog(\Exception $oEx)
-    {
-        if (!($oEx instanceof \OxidEsales\Eshop\Core\Exception\StandardException)) {
-            $oOxException = oxNew(\OxidEsales\Eshop\Core\Exception\StandardException::class);
-            $oOxException->setMessage($oEx->getMessage());
-            $oOxException->debugOut();
-        } else {
-            $oEx->debugOut();
-        }
     }
 
     /**